WIP limit sections in subtasks
The DOM structure for subtasks is completely different (and much
older), so needs different code to handle.
